Use of stinfosys in progres.
git-svn-id: https://svn.met.no/kvoss/kvbufferd/trunk@1932 422d51e2-5941-41ac-8e65-d5b5931239be

+++ src/tblStInfoSysStation.cc
@@ -65,9 +65,9 @@ set(const dnmi::db::DRow &r_)
          if(*it=="stationid"){
             stationid_ = atoi( buf.c_str() );
          }else if(*it=="lat"){
-            lat_ = atoi( buf.c_str() );
+            lat_ = static_cast<float>( atof( buf.c_str() ) );
          }else if(*it=="lon"){
-            lon_ = atoi( buf.c_str() );
+            lon_ = static_cast<float>( atof( buf.c_str() ) );
          }else if(*it=="hs"){
             hs_ = atoi( buf.c_str() );
          }else if(*it=="hv"){
